+++ src/nodejs/openrpg.d.ts
@@ -8,8 +8,21 @@ declare module 'openrpg-libs' {
     export class Die {
         public constructor(max: number);
 
+        /**
+         * @desc randomly generate a integers between 1 and _MAX, where MAX is the value
+         * that was passed to the Die when it was constructed. The result is generated
+         * by using a random_device to seed a Mersenne Twister engine that generates 
+         * uniformly distributed integers between 1 and _MAX.
+         * between 1 and _MAX
+         * @return number - a pesudo random integers between 1 and _MAX
+         */
         public roll(): number;
 
+        /**
+         * @desc returns the number of sides this dice has, this is the same as
+         * the max value that could potentially be rolled.
+         * @return number - the max value that could be returned by this die
+         */
         public MAX(): number;
     }
 
@@ -26,11 +39,26 @@ declare module 'openrpg-libs' {
 
         /**
          * @desc parses the parse_node tree and returns the end result of the expression
-         * @return int - the end result of the expression
+         * @return number - the end result of the expression
          */
         public parse_expression(): number;
 
+        /**
+         * @desc outputs an error with ERROR_CODE if there
+         *     would be an addition overflow
+         * @param number op1 - an integer to be added
+         * @param number op2 - an integer to be added
+         * @return number - op1 + op2
+         */
         public checked_sum(op1: number, op2: number): number;
+
+        /**
+         * @desc outputs an error with ERROR_CODE if there
+         *     would be a multiplication overflow
+         * @param number op1 - an integer to be multiplied
+         * @param number op2 - an integer to be multiply by
+         * @return number - op1 * op2
+         */
         public checked_multiplication(op1: number, op2: number): number;
 
         /**